requirements.txt
Django==1.10b1
mysqlclient==1.3.7
pillow==3.2.0
sorl-thumbnail==12.4.a1
django-ckeditor==5.0.3
-e git+https://github.com/and-nothing-else/django-admin-tools.git@d952cba367bf3d75cf75a7e73d1125fce57c22bd#egg=django_admin_tools

